updated JS-Composer-Salient

+++ wp-content/plugins/js_composer_salient/include/classes/editors/popups/class-vc-templates-panel-editor.php
@@ -7,7 +7,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Class Vc_Templates_Panel_Editor
  * @since 4.4
  */
-class Vc_Templates_Panel_Editor implements Vc_Render {
+class Vc_Templates_Panel_Editor {
 	/**
 	 * @since 4.4
 	 * @var string
@@ -34,16 +34,16 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		}
 		$this->initialized = true;
 		add_filter( 'vc_load_default_templates_welcome_block', array(
-			&$this,
+			$this,
 			'loadDefaultTemplatesLimit',
 		) );
 
 		add_filter( 'vc_templates_render_category', array(
-			&$this,
+			$this,
 			'renderTemplateBlock',
 		), 10 );
 		add_filter( 'vc_templates_render_template', array(
-			&$this,
+			$this,
 			'renderTemplateWindow',
 		), 10, 2 );
 
@@ -55,30 +55,30 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		 *  'vc_delete_template' -> deleting template by index
 		 */
 		add_action( 'wp_ajax_vc_save_template', array(
-			&$this,
+			$this,
 			'save',
 		) );
 		add_action( 'wp_ajax_vc_backend_load_template', array(
-			&$this,
+			$this,
 			'renderBackendTemplate',
 		) );
 		add_action( 'wp_ajax_vc_frontend_load_template', array(
-			&$this,
+			$this,
 			'renderFrontendTemplate',
 		) );
 		add_action( 'wp_ajax_vc_load_template_preview', array(
-			&$this,
+			$this,
 			'renderTemplatePreview',
 		) );
 		add_action( 'wp_ajax_vc_delete_template', array(
-			&$this,
+			$this,
 			'delete',
 		) );
 
-/*		add_action( 'vc-render-templates-preview-template', array(
-			&$this,
-			'addScriptsToTemplatePreview',
-		) );*/
+		/*		add_action( 'vc-render-templates-preview-template', array(
+					$this,
+					'addScriptsToTemplatePreview',
+				) );*/
 
 	}
 
@@ -90,16 +90,12 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		if ( 'my_templates' === $category['category'] ) {
 			$category['output'] = '';
 
-			if ( vc_user_access()
-				->part( 'templates' )
-				->checkStateAny( true, null )
-				->get()
-			) {
+			if ( vc_user_access()->part( 'templates' )->checkStateAny( true, null )->get() ) {
 				$category['output'] .= '
 				<div class="vc_column vc_col-sm-12" data-vc-hide-on-search="true">
 					<div class="vc_element_label">' . esc_html( __( 'Save current layout as a template', 'js_composer' ) ) . '</div>
 					<div class="vc_input-group">
-						<input name="padding" data-js-element="vc-templates-input" class="vc_form-control wpb-textinput vc_panel-templates-name" type="text" value="" placeholder="' . esc_attr( 'Template name', 'js_composer' ) . '" data-vc-disable-empty="#vc_ui-save-template-btn">
+						<input name="padding" data-js-element="vc-templates-input" class="vc_form-control wpb-textinput vc_panel-templates-name" type="text" value="" placeholder="' . esc_attr__( 'Template name', 'js_composer' ) . '" data-vc-disable-empty="#vc_ui-save-template-btn">
 						<span class="vc_input-group-btn">
 							<button class="vc_general vc_ui-button vc_ui-button-size-sm vc_ui-button-action vc_ui-button-shape-rounded vc_template-save-btn" id="vc_ui-save-template-btn" disabled data-vc-ui-element="button-save">' . esc_html( __( 'Save Template', 'js_composer' ) ) . '</button>
 						</span>
@@ -189,18 +185,14 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		$template_id = esc_attr( $template_data['unique_id'] );
 		$template_id_hash = md5( $template_id ); // needed for jquery target for TTA
 		$template_name = esc_html( $template_name );
-		$preview_template_title = esc_attr( 'Preview template', 'js_composer' );
-		$add_template_title = esc_attr( 'Add template', 'js_composer' );
+		$preview_template_title = esc_attr__( 'Preview template', 'js_composer' );
+		$add_template_title = esc_attr__( 'Add template', 'js_composer' );
 		$deleteControlRender = '';
-		if ( vc_user_access()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->get()
-		) {
-			$delete_template_title = esc_attr( 'Delete template', 'js_composer' );
+		if ( vc_user_access()->part( 'templates' )->checkStateAny( true, null )->get() ) {
+			$delete_template_title = esc_attr__( 'Delete template', 'js_composer' );
 			$deleteControlRender = <<<DATA
 				<button type="button" class="vc_general vc_ui-control-button" data-vc-ui-delete="template-title" title="$delete_template_title">
-					<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-trash-dark"></i>
+					<i class="vc-composer-icon vc-c-icon-delete_empty"></i>
 				</button>
 DATA;
 		}
@@ -211,12 +203,12 @@ DATA;
 			<div class="vc_ui-list-bar-item-actions">
 				<button type="button" class="vc_general vc_ui-control-button" title="$add_template_title"
 					 	data-template-handler="">
-					<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-add-dark"></i>
+					<i class="vc-composer-icon vc-c-icon-add"></i>
 				</button>
 				$deleteControlRender
 				<button type="button" class="vc_general vc_ui-control-button" title="$preview_template_title"
 					data-vc-container=".vc_ui-list-bar" data-vc-preview-handler data-vc-target="[data-template_id_hash=$template_id_hash]">
-					<i class="vc_ui-icon-pixel vc_ui-preview-icon"></i>
+					<i class="vc-composer-icon vc-c-icon-arrow_drop_down"></i>
 				</button>
 			</div>
 HTML;
@@ -237,8 +229,8 @@ HTML;
 		$template_id = esc_attr( $template_data['unique_id'] );
 		$template_id_hash = md5( $template_id ); // needed for jquery target for TTA
 		$template_name = esc_html( $template_name );
-		$preview_template_title = esc_attr( 'Preview template', 'js_composer' );
-		$add_template_title = esc_attr( 'Add template', 'js_composer' );
+		$preview_template_title = esc_attr__( 'Preview template', 'js_composer' );
+		$add_template_title = esc_attr__( 'Add template', 'js_composer' );
 
 		echo <<<HTML
 		<button type="button" class="vc_ui-list-bar-item-trigger" title="$add_template_title"
@@ -247,11 +239,11 @@ HTML;
 		<div class="vc_ui-list-bar-item-actions">
 			<button type="button" class="vc_general vc_ui-control-button" title="$add_template_title"
 					data-template-handler="">
-				<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-add-dark"></i>
+				<i class="vc-composer-icon vc-c-icon-add"></i>
 			</button>
 			<button type="button" class="vc_general vc_ui-control-button" title="$preview_template_title"
 				data-vc-container=".vc_ui-list-bar" data-vc-preview-handler data-vc-target="[data-template_id_hash=$template_id_hash]">
-				<i class="vc_ui-icon-pixel vc_ui-preview-icon"></i>
+				<i class="vc-composer-icon vc-c-icon-arrow_drop_down"></i>
 			</button>
 		</div>
 HTML;
@@ -264,23 +256,16 @@ HTML;
 	 * vc_filter: vc_templates_render_frontend_template - called when unknown template received to render in frontend.
 	 */
 	function renderFrontendTemplate() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		add_filter( 'vc_frontend_template_the_content', array(
-			&$this,
+			$this,
 			'frontendDoTemplatesShortcodes',
 		) );
 		$template_id = vc_post_param( 'template_unique_id' );
 		$template_type = vc_post_param( 'template_type' );
 		add_action( 'wp_print_scripts', array(
-			&$this,
+			$this,
 			'addFrontendTemplatesShortcodesCustomCss',
 		) );
 
@@ -325,19 +310,6 @@ HTML;
 	}
 
 	/**
-	 * @since 4.4
-	 * @deprecated 4.7
-	 *
-	 * vc_filter: vc_templates_render_template - hook to override singe template rendering in panel window
-	 */
-	public function render() {
-		_deprecated_function( '\Vc_Templates_Panel_Editor::render', '4.7 (will be removed in 4.11)', '\Vc_Templates_Panel_Editor::renderUITemplate' );
-		vc_include_template( 'editors/popups/panel_templates.tpl.php', array(
-			'box' => $this,
-		) );
-	}
-
-	/**
 	 * @since 4.7
 	 */
 	public function renderUITemplate() {
@@ -352,14 +324,7 @@ HTML;
 	 * @since 4.4
 	 */
 	public function save() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->checkStateAny( true, null )->validateDie();
 
 		$template_name = vc_post_param( 'template_name' );
 		$template = vc_post_param( 'template' );
@@ -401,14 +366,7 @@ HTML;
 	 * vc_filter: vc_templates_render_backend_template - called when unknown template requested to render in backend
 	 */
 	public function renderBackendTemplate() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		$template_id = vc_post_param( 'template_unique_id' );
 		$template_type = vc_post_param( 'template_type' );
@@ -443,17 +401,10 @@ HTML;
 	 * @since 4.8
 	 */
 	public function renderTemplatePreview() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( array(
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( array(
 				'edit_post',
 				(int) vc_request_param( 'post_id' ),
-			) )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+			) )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		$template_id = vc_request_param( 'template_unique_id' );
 		$template_type = vc_request_param( 'template_type' );
@@ -487,6 +438,7 @@ HTML;
 		die();
 
 	}
+
 	public function registerPreviewScripts() {
 		visual_composer()->registerAdminJavascript();
 		visual_composer()->registerAdminCss();
@@ -496,6 +448,7 @@ HTML;
 			'vc-backend-min-js',
 		), WPB_VC_VERSION, true );
 	}
+
 	/**
 	 * Enqueue required scripts for template preview
 	 * @since 4.8
@@ -510,29 +463,28 @@ HTML;
 	 * @since 4.4
 	 */
 	public function delete() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->checkStateAny( true, null )->validateDie();
 
 		$template_id = vc_post_param( 'template_id' );
+		$template_type = vc_post_param( 'template_type' );
 
 		if ( ! isset( $template_id ) || '' === $template_id ) {
 			die( 'Error: Vc_Templates_Panel_Editor::delete:1' );
 		}
+		if ( 'my_templates' === $template_type ) {
 
-		$saved_templates = get_option( $this->option_name );
-		unset( $saved_templates[ $template_id ] );
-		if ( count( $saved_templates ) > 0 ) {
-			update_option( $this->option_name, $saved_templates );
+			$saved_templates = get_option( $this->option_name );
+			unset( $saved_templates[ $template_id ] );
+			if ( count( $saved_templates ) > 0 ) {
+				update_option( $this->option_name, $saved_templates );
+			} else {
+				delete_option( $this->option_name );
+			}
+			wp_send_json_success();
 		} else {
-			delete_option( $this->option_name );
+			do_action( 'vc_templates_delete_templates', $template_id, $template_type );
 		}
-		die();
+		wp_send_json_error();
 	}
 
 	/**
@@ -583,8 +535,10 @@ HTML;
 			// this has only 'name' and 'template' key  and index 'key' is template id.
 			$arr_category = array(
 				'category' => 'my_templates',
+				/*nectar addition*/
 				'category_name' => __( 'My Templates', 'js_composer' ),
-				'category_description' => __( 'Append previously saved template to the current layout.', 'js_composer' ),
+				'category_description' => __( 'Append previously saved templates to the current layout.', 'js_composer' ),
+				/*nectar addition end*/
 				'category_weight' => 10,
 			);
 			$category_templates = array();
@@ -595,6 +549,12 @@ HTML;
 						'name' => $template_data['name'],
 						'type' => 'my_templates',
 						// for rendering in backend/frontend with ajax
+						/*nectar addition*/
+						'image' => isset( $template_data['image_path'] ) ? $template_data['image_path'] : false,
+						'cat_display_name' => isset( $template_data['cat_display_name'] ) ? $template_data['cat_display_name'] : '',
+						/*nectar addition end*/
+						// preview image
+						'custom_class' => isset( $template_data['custom_class'] ) ? $template_data['custom_class'] : false,
 					);
 				}
 			}
@@ -607,13 +567,18 @@ HTML;
 		// this has 'name', 'image_path', 'custom_class', and 'content' as template data
 		if ( ! empty( $default_templates ) ) {
 			$arr_category = array(
+				/*nectar addition*/
 				'category' => 'default_templates',
-				'category_name' => __( 'Default Templates', 'js_composer' ),
-				'category_description' => __( 'Append default template to the current layout.', 'js_composer' ),
+				'category_name' => __( 'Salient Studio', 'js_composer' ),
+				'category_description' => __( 'Append predefined Salient templates to the current layout', 'js_composer' ),
 				'category_weight' => 11,
+				/*nectar addition end*/
 			);
 			$category_templates = array();
 			foreach ( $default_templates as $template_id => $template_data ) {
+				if ( isset( $template_data['disabled'] ) && $template_data['disabled'] ) {
+					continue;
+				}
 				$category_templates[] = array(
 					'unique_id' => $template_id,
 					'name' => $template_data['name'],
@@ -624,8 +589,10 @@ HTML;
 					'custom_class' => isset( $template_data['custom_class'] ) ? $template_data['custom_class'] : false,
 				);
 			}
-			$arr_category['templates'] = $category_templates;
-			$data[] = $arr_category;
+			if ( ! empty( $category_templates ) ) {
+				$arr_category['templates'] = $category_templates;
+				$data[] = $arr_category;
+			}
 		}
 
 		// To get any other 3rd "Custom template" - do this by hook filter 'vc_get_all_templates'
@@ -708,7 +675,6 @@ HTML;
 			if ( ! is_array( $this->default_templates ) ) {
 				$this->default_templates = array();
 			}
-
 			$this->default_templates[] = $data;
 
 			return true;
@@ -749,7 +715,7 @@ HTML;
 	public function sortTemplatesByCategories( array $data ) {
 		$buffer = $data;
 		uasort( $buffer, array(
-			&$this,
+			$this,
 			'cmpCategory',
 		) );
 
@@ -766,7 +732,7 @@ HTML;
 	public function sortTemplatesByNameWeight( array $data ) {
 		$buffer = $data;
 		uasort( $buffer, array(
-			&$this,
+			$this,
 			'cmpNameWeight',
 		) );
 
@@ -864,7 +830,6 @@ HTML;
 	 * @todo move to autoload or else some where.
 	 * @since 4.4.3
 	 *
-	 * @return string
 	 */
 	public function addFrontendTemplatesShortcodesCustomCss() {
 		$output = $shortcodes_custom_css = '';
@@ -890,6 +855,9 @@ HTML;
 		$template_name_lower = esc_attr( vc_slugify( $template_name ) );
 		$template_type = esc_attr( isset( $template['type'] ) ? $template['type'] : 'custom' );
 		$custom_class = esc_attr( isset( $template['custom_class'] ) ? $template['custom_class'] : '' );
+		/*nectar addition - adding in preview img*/
+		$preview_img = esc_attr( isset( $template['image'] ) &&  $template['image'] != '' ? $template['image'] : get_template_directory_uri() .'/nectar/nectar-vc-addons/img/templates/no-img.jpg'  );
+		$cat_display_name = esc_attr( isset( $template['cat_display_name'] ) ? $template['cat_display_name'] : '' );
 
 		$output = <<<HTML
 					<div class="vc_ui-template vc_templates-template-type-$template_type $custom_class"
@@ -902,6 +870,8 @@ HTML;
 						data-vc-content=".vc_ui-template-content">
 						<div class="vc_ui-list-bar-item">
 HTML;
+
+	if(!empty($preview_img) && $template_type == 'default_templates') $output .= '<div class="img-wrap"><img src="'.$preview_img.'" alt="'.$name.'" width="300" height="200" /></div><div class="display_cat">'.$cat_display_name .'</div>';
 		$output .= apply_filters( 'vc_templates_render_template', $name, $template );
 		$output .= <<<HTML
 						</div>
@@ -909,6 +879,8 @@ HTML;
 						</div>
 					</div>
 HTML;
+	/*nectar addition end*/
+
 
 		return $output;
 	}
